--- src/ducks/Studio/Chart/MetricSettings/utils.js
@@ -1,8 +1,0 @@
-export function getMetricSetting (MetricSettingMap, metric) {
-  const metricSetting = MetricSettingMap.get(metric)
-  if (metricSetting) return metricSetting
-
-  const newMetricSetting = {}
-  MetricSettingMap.set(metric, newMetricSetting)
-  return newMetricSetting
-}
