+++ src/ducks/Studio/index.js
@@ -104,8 +104,7 @@ export const Studio = ({
     if (
       widget.Widget !== HolderDistributionWidget &&
       widget.Widget !== HolderDistributionCombinedBalanceWidget &&
-      metrics.length === 0 &&
-      widget.comparables.length === 0
+      metrics.length === 0
     ) {
       deleteWidget(widget)
     } else {
