+++ src/ducks/Studio/utils.js
@@ -59,3 +59,12 @@ export function deduceItems (items, item) {
 
   return [...newItems]
 }
+
+export function getMetricSetting (MetricSettingMap, metric) {
+  const metricSetting = MetricSettingMap.get(metric)
+  if (metricSetting) return metricSetting
+
+  const newMetricSetting = {}
+  MetricSettingMap.set(metric, newMetricSetting)
+  return newMetricSetting
+}
